refactor(hydration): move hydration query into feature
