+++ src/Arius.Cli.Tests/CliTests.cs
@@ -1,4 +1,5 @@
 using Arius.Core.Features.Archive;
+using Arius.Core.Features.Hydration;
 using Arius.Core.Features.List;
 using Arius.Core.Features.Restore;
 using Arius.Core.Shared.Storage;
