+++ src/Arius.Cli.Tests/DependencyInjectionTests.cs
@@ -1,5 +1,6 @@
 using Arius.Core;
 using Arius.Core.Features.Archive;
+using Arius.Core.Features.Hydration;
 using Arius.Core.Features.List;
 using Arius.Core.Features.Restore;
 using Arius.Core.Shared.Storage;
