+++ src/Arius.Core.Tests/List/FileHydrationStatusResolverTests.cs
@@ -1,3 +1,4 @@
+using Arius.Core.Features.Hydration;
 using Arius.Core.Features.List;
 using Arius.Core.Shared.Storage;
 using Shouldly;
