+++ src/Arius.Core/Features/Hydration/HydrationModels.cs
@@ -1,8 +1,9 @@
+using Arius.Core.Features.List;
 using Arius.Core.Shared.Storage;
 using Mediator;
 using System.Collections.Generic;
 
-namespace Arius.Core.Features.List;
+namespace Arius.Core.Features.Hydration;
 
 public enum FileHydrationStatus
 {
