+++ src/Arius.Core/Features/Restore/RestorePipelineHandler.cs
@@ -1,4 +1,5 @@
 using Arius.Core.Shared.ChunkIndex;
+using Arius.Core.Features.Hydration;
 using Arius.Core.Features.List;
 using Arius.Core.Shared.Encryption;
 using Arius.Core.Shared.FileTree;
