+++ src/Arius.Explorer/RepositoryExplorer/RepositoryExplorerViewModel.cs
@@ -1,3 +1,4 @@
+using Arius.Core.Features.Hydration;
 using Arius.Core.Features.List;
 using Arius.Core.Features.Restore;
 using Arius.Explorer.Infrastructure;
